# swagger/swaggerProOrg1.json
{
  "swagger": "2.0",
  "info": {
    "description": "13 jan 2020",
    "version": "2",
    "title": "Sampran"
  },
  "host": "206.189.152.78:7777",
...
  "schemes": [
    "http"
  ],
  "paths": {
    "/RegisterUser": {
      "post": {
        "tags": [
          "registerUser"
        ],
        "summary": "register a signgle Blockchain user",
        "description": "Register Blockchain user for Organisation Store in MongoDB",
        "operationId": "register for some users",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "bc_user",
            "in": "formData",
            "description": "Blockchain user",
            "required": true,
            "type": "string",
            "default": "123456"
          },
          {
            "name": "OrgDepartment",
            "in": "formData",
            "description": "Department in Organisation",
            "required": true,
            "type": "string",
            "default": "org1.department1"
          },
          {
            "name": "apikey",
            "in": "header",
            "description": "API Key an authorization for api server send with header request",
            "required": true,
            "type": "string",
            "default": "558344f429d66e42b53ff6449710cea0f79211c2163d3d92b4180eaee3b6afd1"
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/registerUser201"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/badRequest400"
            }
          },
          "401": {
            "description": "unAuthenticate",
            "schema": {
              "$ref": "#/definitions/unAuthenticate401"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/registerUser500"
            }
          }
        }
      }
    },
    "/registerConsumer": {
      "post": {
        "tags": [
          "Consumer"
        ],
        "summary": "Create a single Document Consumer",
        "description": "Create Document Consumer in Blockchain Network",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "request body registerConsumer",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ConsumerInfo"
            }
          },
          {
            "name": "apikey",
            "in": "header",
            "description": "API Key an authorization for api server send with header request",
            "required": true,
            "type": "string",
            "default": "558344f429d66e42b53ff6449710cea0f79211c2163d3d92b4180eaee3b6afd1"
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/registerConsumer201"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/badRequest400"
            }
          },
          "401": {
            "description": "unAuthenticate",
            "schema": {
              "$ref": "#/definitions/unAuthenticate401"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/registerConsumer500"
            }
          }
        }
      }
    },
    "/addPoint": {
      "put": {
        "tags": [
          "Consumer"
        ],
        "summary": "Update Point a single Document Consumer",
        "description": "Create Document Consumer in Blockchain Network",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "request body IssueGarden",
            "required": true,
            "schema": {
              "$ref": "#/definitions/addPointInfo"
            }
          },
          {
            "name": "apikey",
            "in": "header",
            "description": "API Key an authorization for api server send with header request",
            "required": true,
            "type": "string",
            "default": "558344f429d66e42b53ff6449710cea0f79211c2163d3d92b4180eaee3b6afd1"
          }
        ],
        "responses": {
          "200": {
            "description": "Updated",
            "schema": {
              "$ref": "#/definitions/addPoint200"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/badRequest400"
            }
          },
          "401": {
            "description": "unAuthenticate",
            "schema": {
              "$ref": "#/definitions/unAuthenticate401"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/addPoint500"
            }
          }
        }
      }
    },
    "/usePoint": {
      "put": {
        "tags": [
          "Consumer"
        ],
        "summary": "Update Point a single Document Consumer",
        "description": "Create Document Consumer in Blockchain Network",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "request body IssueGarden",
            "required": true,
            "schema": {
              "$ref": "#/definitions/usePointInfo"
            }
          },
          {
            "name": "apikey",
            "in": "header",
            "description": "API Key an authorization for api server send with header request",
            "required": true,
            "type": "string",
            "default": "558344f429d66e42b53ff6449710cea0f79211c2163d3d92b4180eaee3b6afd1"
          }
        ],
        "responses": {
          "200": {
            "description": "Updated",
            "schema": {
              "$ref": "#/definitions/usePoint200"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/badRequest400"
            }
          },
          "401": {
            "description": "unAuthenticate",
            "schema": {
              "$ref": "#/definitions/unAuthenticate401"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/usePoint500"
            }
          }
        }
      }
    },
    "/queryConsumer": {
      "post": {
        "tags": [
          "Consumer"
        ],
        "summary": "get a single Document Consumer",
        "description": "get Document Consumer form Blockchain Network",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "request body QueryConsumer",
            "required": true,
            "schema": {
              "$ref": "#/definitions/queryConsumerInfo"
            }
          },
          {
            "name": "apikey",
            "in": "header",
            "description": "API Key an authorization for api server send with header request",
            "required": true,
            "type": "string",
            "default": "558344f429d66e42b53ff6449710cea0f79211c2163d3d92b4180eaee3b6afd1"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/queryConsumer200"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/badRequest400"
            }
          },
          "401": {
            "description": "unAuthenticate",
            "schema": {
              "$ref": "#/definitions/unAuthenticate401"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/queryConsumer500"
            }
          }
        }
      }
    },
    "/IssueGarden": {
      "post": {
        "tags": [
          "issueDocument"
        ],
        "summary": "Create a signgle Document Garden",
        "description": "Create Document Garden store in Blockchain Network",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "request body IssueGarden",
            "required": true,
            "schema": {
              "$ref": "#/definitions/issueGardenInfo"
            }
          },
          {
            "name": "apikey",
            "in": "header",
            "description": "API Key an authorization for api server send with header request",
            "required": true,
            "type": "string",
            "default": "558344f429d66e42b53ff6449710cea0f79211c2163d3d92b4180eaee3b6afd1"
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/issueGarden201"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/badRequest400"
            }
          },
          "401": {
            "description": "unAuthenticate",
            "schema": {
              "$ref": "#/definitions/unAuthenticate401"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/issueGarden500"
            }
          }
        }
      }
    },
    "/IssueProduct": {
      "post": {
        "tags": [
          "E-Commerce"
        ],
        "summary": "Create a signgle Document Product",
        "description": "Create Document Product in Blockchain Network",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "request body IssueProduct",
            "required": true,
            "schema": {
              "$ref": "#/definitions/issueProductInfo"
            }
          },
          {
            "name": "apikey",
            "in": "header",
            "description": "API Key an authorization for api server send with header request",
            "required": true,
            "type": "string",
            "default": "558344f429d66e42b53ff6449710cea0f79211c2163d3d92b4180eaee3b6afd1"
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/issueProduct201"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/badRequest400"
            }
          },
          "401": {
            "description": "unAuthenticate",
            "schema": {
              "$ref": "#/definitions/unAuthenticate401"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/issueGarden500"
            }
          }
        }
      }
    },
...
},
  "definitions": {
    "unAuthenticate401": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "number",
          "example": 401
        },
        "message": {
          "type": "string",
          "example": "Failed to Authenticate API KEY"
        }
      }
    },
    "badRequest400": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "number",
          "example": 400
        },
        "message": {
          "type": "string",
          "example": "<function_name>.bc_user is required"
        }
      }
    },
    "registerUser201": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "number",
          "example": 201
        },
        "message": {
          "type": "string",
          "example": "Successfully registered blockchain user 1234564562611 and insert it into mongoDB"
        }
      }
    },
    "registerUser500": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "number",
          "example": 500
        },
        "message": {
          "type": "string",
          "example": "[Blockchain.registerUser] [Error] An identity for the user 123456456 already exists in the MongoDB"
        }
      }
    },
...
    "registerConsumer201": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "number",
          "example": 201
        },
        "message": {
          "type": "string",
          "example": "Request for [registerConsumer] Sampran|6b86b273ff34fce19d6b804eff5a3f5747ada4eaa22f1d49c01e52ddb7875b4b : Transaction has been submitted"
        }
      }
    },
...
    "ConsumerInfo": {
      "type": "object",
      "required": [
        "bc_user"
      ],
      "properties": {
        "bc_user": {
          "type": "string",
          "example": "123456"
        },
        "app_user": {
          "type": "string",
          "example": "Ohmsm0150"
        },
        "Platform": {
          "type": "string",
          "example": "Sampran"
        }
      },
      "title": "ConsumerInfo"
    },
...
  }
}